+++ src/MEventAction.cc
@@ -11,8 +11,6 @@
 // mlibrary
 #include "frequencySyncSignal.h"
 
-#include <sys/types.h>
-#include <dirent.h>
 #include <iostream>
 using namespace std;
 
@@ -166,61 +164,6 @@ MEventAction::MEventAction(goptions opts, map<string, double> gpars)
 	}
     }
 
-  // Set up to read saved RNGs
-
-  arg = gemcOpt.optMap["RERUN_SELECTED"].args;
-  cout << ">>>>>> setup RERUN_SELECTED: [" << arg << "]" << endl;
-  if (arg == "" || arg == "no")
-    rsp.enabled = false;
-  else
-    {
-      vector<string> values;
-      string units;
-      values       = get_info(gemcOpt.optMap["RERUN_SELECTED"].args, string(",\""));
-      if (values.size() <= 2)
-	{
-	  rsp.enabled = true;
-	  rsp.run = get_number (values[0]);
-	  if (values.size() == 1)
-	    rsp.dir = "./";
-	  else
-	    {
-	      rsp.dir = values[1];
-	      if (rsp.dir[rsp.dir.size()-1] != '/' ) rsp.dir += "/";
-#ifdef WIN32
-	      std::replace(rsp.dir.begin(), rsp.dir.end(),'/','\\');
-#endif
-	    }
-
-	  DIR* dirp = opendir(rsp.dir.c_str());
-	  struct dirent * dp;
-	  while ((dp = readdir(dirp)) != NULL)
-	    {
-	      string dname (dp->d_name);
-	      size_t rpos = dname.find ("run");
-	      size_t epos = dname.find ("evt");
-	      size_t rnpos = dname.find (".rndm");
-	      if (rpos == string::npos || epos == string::npos
-		  || rnpos != dname.size()-5)
-		continue;
-
-	      unsigned rstring = get_number (dname.substr (rpos+3, epos-rpos-3));
-	      if (rstring == rsp.run)
-		{
-		  string estring = dname.substr (epos+3, rnpos-epos-3);
-		  rsp.events.push_back (atoi (estring.c_str()));
-		  cout << dname << " " << epos << " " << string::npos << " " << estring << endl;
-		}
-	    }
-	  closedir(dirp);
-	  std::sort (rsp.events.begin(), rsp.events.end());
-	  cout << "sorted:" << endl;
-	  for (vector<unsigned>::const_iterator i = rsp.events.begin(); i != rsp.events.end(); ++i)
-	    cout << (*i) << endl;
-	  rsp.currentevent = -1;
-	}
-    }
-  
 }
 
 MEventAction::~MEventAction()
@@ -240,22 +183,6 @@ void MEventAction::BeginOfEventAction(const G4Event* evt)
 
 	rw.getRunNumber(evtN);
 	bgMap.clear();
-
-	// Seed RNG if required
-
-	if (rsp.enabled)
-	  {
-	    ++rsp.currentevent;
-	    if (rsp.currentevent < int(rsp.events.size()))
-	      cout << "Here I would set up event " << rsp.events[rsp.currentevent] << endl;
-	    else
-	      {
-		G4RunManager *runManager = G4RunManager::GetRunManager();;
-		runManager->AbortRun();
-		cout << " No more events to rerun." << endl;
-		return;
-	      }
-	  }
 
 	if(evtN%Modulo == 0 ) {
 		cout << hd_msg << " Begin of event " << evtN << "  Run Number: " << rw.runNo;
@@ -290,9 +217,13 @@ void MEventAction::BeginOfEventAction(const G4Event* evt)
 
 void MEventAction::EndOfEventAction(const G4Event* evt)
 {
-	if(gen_action->isFileOpen() == false) {
-		return;
-	}
+  cout << "eea " << evtN << endl;
+
+  if ((gen_action->isFileOpen() == false) ||
+      (gen_action->doneRerun() == true))
+      return;
+
+  cout << "proceeding" << endl;
 
 	MHitCollection* MHC;
 	int nhits;
